musl: Update to musl 1.2.0.

+++ musl/src/mman/shm_open.c
@@ -5,27 +5,39 @@
 #include <string.h>
 #include <limits.h>
 #include <pthread.h>
-#include "syscall.h"
 
-#define SHM_PATH_PREFIX		"/dev/shm"
-#define SHM_PATH_PREFIX_SIZE	sizeof(SHM_PATH_PREFIX)
+char *__shm_mapname(const char *name, char *buf)
+{
+	char *p;
+	while (*name == '/') name++;
+	if (*(p = __strchrnul(name, '/')) || p==name ||
+	    (p-name <= 2 && name[0]=='.' && p[-1]=='.')) {
+		errno = EINVAL;
+		return 0;
+	}
+	if (p-name > NAME_MAX) {
+		errno = ENAMETOOLONG;
+		return 0;
+	}
+	memcpy(buf, "/dev/shm/", 9);
+	memcpy(buf+9, name, p-name+1);
+	return buf;
+}
 
 int shm_open(const char *name, int flag, mode_t mode)
 {
-	char buf[NAME_MAX + SHM_PATH_PREFIX_SIZE + 2] = {0};
-	strcpy(buf, SHM_PATH_PREFIX);
-	memcpy(buf + SHM_PATH_PREFIX_SIZE, name, strlen(name));
-
-	int fd = open(buf, flag | O_CLOEXEC, mode);
-
+	int cs;
+	char buf[NAME_MAX+10];
+	if (!(name = __shm_mapname(name, buf))) return -1;
+	pthread_setcancelstate(PTHREAD_CANCEL_DISABLE, &cs);
+	int fd = open(name, flag|O_NOFOLLOW|O_CLOEXEC|O_NONBLOCK, mode);
+	pthread_setcancelstate(cs, 0);
 	return fd;
 }
 
 int shm_unlink(const char *name)
 {
-	char buf[NAME_MAX + SHM_PATH_PREFIX_SIZE + 2] = {0};
-	strcpy(buf, SHM_PATH_PREFIX);
-	memcpy(buf + SHM_PATH_PREFIX_SIZE, name, strlen(name));
-
+	char buf[NAME_MAX+10];
+	if (!(name = __shm_mapname(name, buf))) return -1;
 	return unlink(name);
 }
